requirements.txt
adjustText==0.7.3
aequitas==0.42.0
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
alembic==1.7.7
altair==4.1.0
anyio==3.6.1
argcmdr==0.7.0
argcomplete==1.9.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
awscli==1.25.16
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.22.4
botocore==1.25.13
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
csvkit==1.0.7
cycler==0.11.0
dbfread==2.0.7
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Dickens==1.0.1
docutils==0.16
dominate==2.7.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
fonttools==4.33.3
fsspec==2022.7.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.20
greenlet==1.1.2
html5lib==1.1
httplib2==0.20.4
humanfriendly==10.0
idna==3.3
inflection==0.5.1
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.9
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.3
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.3
leather==0.3.4
Mako==1.2.1
markdown2==2.3.5
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
millify==0.1.1
mistune==0.8.4
nbclassic==0.4.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.3
ohio==0.5.0
olefile==0.46
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parsedatetime==2.4
parso==0.8.3
Pebble==4.6.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
plumbum==1.6.4
prettytable==0.7.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
PyPDF2==2.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytimeparse==1.1.8
pytorch-tabnet==3.1.1
pytz==2022.1
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
reportlab==3.6.11
requests==2.27.1
retrying==1.3.3
rsa==4.7.2
s3fs==0.4.2
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
signalled-timeout==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.18
sqlalchemy-postgres-copy==0.5.0
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.2
terminado==0.15.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.12.0
torch==1.12.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
triage==5.2.0
typing_extensions==4.3.0
urllib3==1.26.11
verboselogs==1.7
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.1
widgetsnbextension==3.6.0
wrapt==1.14.0
xhtml2pdf==0.2.2
xlrd==2.0.1
